Move all non-command code out of commands.py
